Rewrite MainPanel list
- Other: Change FavCimsCore.RowID to list
